Discard stderr from remote shell
we can't reasonably deal with it as long as we read in blocking mode. This needs asyncio or sth similar.
See also: comments in PR #19 and issue #20

Note, however, that at least in theory this means a remaining issue with get/get_from_archive, which can
theoretically hang forever under some circumstances.